+++ python/neuromation/cli/config.py
@@ -3,9 +3,12 @@ from yarl import URL
 
 from . import rc
 from .defaults import API_URL
+from .formatter import ConfigFormatter
+from .rc import Config
+from .utils import group
 
 
-@click.group()
+@group()
 def config() -> None:
     """Client configuration settings commands."""
 
@@ -14,12 +17,11 @@ def config() -> None:
 @click.argument("url")
 def url(url: str) -> None:
     """
-    Updates settings with provided platform URL.
+    Update settings with provided platform URL.
 
     Examples:
 
-    \b
-        neuro config url https://platform.neuromation.io/api/v1
+    neuro config url https://platform.neuromation.io/api/v1
     """
     rc.ConfigFactory.update_api_url(url)
 
@@ -28,7 +30,7 @@ def url(url: str) -> None:
 @click.argument("file", type=click.Path(exists=True, readable=True, dir_okay=False))
 def id_rsa(file: str) -> None:
     """
-    Updates path to id_rsa file with private key.
+    Update path to id_rsa file with private key.
 
     FILE is being used for accessing remote shell, remote debug.
 
@@ -39,19 +41,29 @@ def id_rsa(file: str) -> None:
 
 
 @config.command()
-def show() -> None:
+@click.pass_obj
+def show(cfg: Config) -> None:
     """
-    Prints current settings.
+    Print current settings.
     """
-    config = rc.ConfigFactory.load()
-    click.echo(config)
+    fmt = ConfigFormatter()
+    click.echo(fmt(cfg))
+
+
+@config.command()
+@click.pass_obj
+def show_token(cfg: Config) -> None:
+    """
+    Print current authorization token.
+    """
+    click.echo(cfg.auth)
 
 
 @config.command()
 @click.argument("token")
 def auth(token: str) -> None:
     """
-    Updates authorization token.
+    Update authorization token.
     """
     # TODO (R Zubairov, 09/13/2018): check token correct
     # connectivity, check with Alex
@@ -69,7 +81,7 @@ def forget() -> None:
     rc.ConfigFactory.forget_auth_token()
 
 
-@click.command()
+@config.command()
 @click.argument("url", required=False, default=API_URL, type=URL)
 def login(url: URL) -> None:
     """
@@ -79,7 +91,7 @@ def login(url: URL) -> None:
     click.echo(f"Logged into {url}")
 
 
-@click.command()
+@config.command()
 def logout() -> None:
     """
     Log out.