+++ python/neuromation/cli/job.py
@@ -19,20 +19,20 @@ from .defaults import (
     JOB_SSH_USER,
 )
 from .formatter import (
+    JobFormatter,
     JobListFormatter,
     JobStatusFormatter,
     JobTelemetryFormatter,
-    OutputFormatter,
 )
 from .rc import Config
 from .ssh_utils import connect_ssh
-from .utils import run_async
+from .utils import alias, group, run_async
 
 
 log = logging.getLogger(__name__)
 
 
-@click.group()
+@group()
 def job() -> None:
     """
     Job operations.
@@ -138,19 +138,17 @@ async def submit(
 
     Examples:
 
-    \b
     # Starts a container pytorch:latest with two paths mounted. Directory /q1/
     # is mounted in read only mode to /qm directory within container.
     # Directory /mod mounted to /mod directory in read-write mode.
     neuro job submit --volume storage:/q1:/qm:ro --volume storage:/mod:/mod:rw \
-    pytorch:latest
+      pytorch:latest
 
-    \b
     # Starts a container pytorch:latest with connection enabled to port 22 and
     # sets PYTHONPATH environment value to /python.
     # Please note that SSH server should be provided by container.
     neuro job submit --env PYTHONPATH=/python --volume \
-    storage:/data/2018q1:/data:ro --ssh 22 pytorch:latest
+      storage:/data/2018q1:/data:ro --ssh 22 pytorch:latest
     """
 
     username = cfg.username
@@ -189,7 +187,7 @@ async def submit(
             description=description,
             env=env_dict,
         )
-        click.echo(OutputFormatter().format_job(job, quiet))
+        click.echo(JobFormatter()(job, quiet))
 
 
 @job.command(context_settings=dict(ignore_unknown_options=True))
@@ -220,7 +218,7 @@ async def exec(
     sys.exit(retcode)
 
 
-@job.command()
+@job.command(deprecated=True)
 @click.argument("id")
 @click.option(
     "--user", help="Container user name", default=JOB_SSH_USER, show_default=True
@@ -231,11 +229,11 @@ async def exec(
 async def ssh(cfg: Config, id: str, user: str, key: str) -> None:
     """
     Starts ssh terminal connected to running job.
+
     Job should be started with SSH support enabled.
 
     Examples:
 
-    \b
     neuro job ssh --user alfa --key ./my_docker_id_rsa job-abc-def-ghk
     """
     git_key = cfg.github_rsa_path
@@ -248,9 +246,9 @@ async def ssh(cfg: Config, id: str, user: str, key: str) -> None:
 @click.argument("id")
 @click.pass_obj
 @run_async
-async def monitor(cfg: Config, id: str) -> None:
+async def logs(cfg: Config, id: str) -> None:
     """
-    Monitor job output stream
+    Fetch the logs of a container.
     """
     timeout = aiohttp.ClientTimeout(
         total=None, connect=None, sock_read=None, sock_connect=30
@@ -261,6 +259,9 @@ async def monitor(cfg: Config, id: str) -> None:
             if not chunk:
                 break
             click.echo(chunk.decode(errors="ignore"), nl=False)
+
+
+job.add_command(alias(logs, "monitor"))
 
 
 @job.command()
@@ -280,15 +281,12 @@ async def monitor(cfg: Config, id: str) -> None:
 @click.option("-q", "--quiet", is_flag=True)
 @click.pass_obj
 @run_async
-async def list(
-    cfg: Config, status: Sequence[str], description: str, quiet: bool
-) -> None:
+async def ls(cfg: Config, status: Sequence[str], description: str, quiet: bool) -> None:
     """
     List all jobs.
 
     Examples:
 
-    \b
     neuro job list --description="my favourite job"
     neuro job list --status=all
     neuro job list -s pending -s running -q
@@ -305,7 +303,10 @@ async def list(
         jobs = await client.jobs.list()
 
     formatter = JobListFormatter(quiet=quiet)
-    click.echo(formatter.format_jobs(jobs, statuses, description))
+    click.echo(formatter(jobs, statuses, description))
+
+
+job.add_command(alias(ls, "list"))
 
 
 @job.command()
@@ -314,11 +315,11 @@ async def list(
 @run_async
 async def status(cfg: Config, id: str) -> None:
     """
-    Display status of a job
+    Display status of a job.
     """
     async with cfg.make_client() as client:
         res = await client.jobs.status(id)
-        click.echo(JobStatusFormatter().format_job_status(res))
+        click.echo(JobStatusFormatter()(res))
 
 
 @job.command()
@@ -327,16 +328,16 @@ async def status(cfg: Config, id: str) -> None:
 @run_async
 async def top(cfg: Config, id: str) -> None:
     """
-    Display real-time job telemetry
+    Display real-time job telemetry.
     """
     formatter = JobTelemetryFormatter()
     async with cfg.make_client() as client:
         print_header = True
         async for res in client.jobs.top(id):
             if print_header:
-                click.echo(formatter.format_header())
+                click.echo(formatter.header())
                 print_header = False
-            line = formatter.format(res)
+            line = formatter(res)
             click.echo(f"\r{line}", nl=False)
 
 
@@ -346,7 +347,7 @@ async def top(cfg: Config, id: str) -> None:
 @run_async
 async def kill(cfg: Config, id: Sequence[str]) -> None:
     """
-    Kill job(s)
+    Kill job(s).
     """
     errors = []
     async with cfg.make_client() as client: